# Conflicts:
#	Service/AutoBidSettingsManager.php
#	apps/default/controllers/ajax.php

+++ Service/AutoBidSettingsManager.php
@@ -9,15 +9,18 @@ use Unilend\core\Loader;
  */
 class AutoBidSettingsManager
 {
-    const AUTO_BID_ON  = 1;
-    const AUTO_BID_OFF = 0;
-
     /** @var BidManager */
     private $oBidManager;
 
+    /**
+     * @var ClientSettingsManager
+     */
+    private $oClientSettingsManager;
+
     public function __construct()
     {
-        $this->oBidManager = Loader::loadService('BidManager');
+        $this->oBidManager            = Loader::loadService('BidManager');
+        $this->oClientSettingsManager = Loader::loadService('ClientSettingsManager');
     }
 
     /**
@@ -30,9 +33,10 @@ class AutoBidSettingsManager
         /** @var \autobid_queue $oAutoBidQueue */
         $oAutoBidQueue = Loader::loadData('autobid_queue');
 
-        if ($this->isQualified($oClient) && $this->oBidManager->canBid($oClient)) {
-            $this->onOff($oClient, self::AUTO_BID_ON);
-
+        if ($this->isQualified($oClient)
+            && $this->oBidManager->canBid($oClient)
+            && $this->oClientSettingsManager->saveClientSetting($oClient, \client_setting_type::TYPE_AUTO_BID_SWITCH, \client_settings::AUTO_BID_ON)) {
+            $this->saveAutoBidSwitchHistory($oClient->id_client, \client_settings::AUTO_BID_ON);
             $oLendersAccount->get($oClient->id_client, 'id_client_owner');
             $oAutoBidQueue->addToQueue($oLendersAccount->id_lender_account, \autobid_queue::TYPE_QUEUE_NEW);
         }
@@ -48,38 +52,11 @@ class AutoBidSettingsManager
         /** @var \autobid_queue $oAutoBidQueue */
         $oAutoBidQueue = Loader::loadData('autobid_queue');
 
-        $this->onOff($oClient, self::AUTO_BID_OFF);
-
-        $oLendersAccount->get($oClient->id_client, 'id_client_owner');
-        $oAutoBidQueue->delete($oLendersAccount->id_lender_account, 'id_lender');
-    }
-
-    /**
-     * @param \clients $oClient
-     * @param          $iAutoBidOnOff
-     */
-    private function onOff(\clients $oClient, $iAutoBidOnOff)
-    {
-        /** @var \client_settings $oClientSettings */
-        $oClientSettings = Loader::loadData('client_settings');
-        /** @var \clients_history_actions $oClientHistoryActions */
-        $oClientHistoryActions = Loader::loadData('clients_history_actions');
-
-        if ($oClientSettings->get($oClient->id_client, 'id_type = ' . \client_setting_type::TYPE_AUTO_BID_SWITCH . ' AND id_client')) {
-            $oClientSettings->value = $iAutoBidOnOff;
-            $oClientSettings->update();
-        } else {
-            $oClientSettings->unsetData();
-            $oClientSettings->id_client = $oClient->id_client;
-            $oClientSettings->id_type   = \client_setting_type::TYPE_AUTO_BID_SWITCH;
-            $oClientSettings->value     = $iAutoBidOnOff;
-            $oClientSettings->create();
+        if ($this->oClientSettingsManager->saveClientSetting($oClient, \client_setting_type::TYPE_AUTO_BID_SWITCH, \client_settings::AUTO_BID_OFF)) {
+            $this->saveAutoBidSwitchHistory($oClient->id_client, \client_settings::AUTO_BID_OFF);
+            $oLendersAccount->get($oClient->id_client, 'id_client_owner');
+            $oAutoBidQueue->delete($oLendersAccount->id_lender_account, 'id_lender');
         }
-        // BO user
-        $iUserId     = isset($_SESSION['user']['id_user']) ? $_SESSION['user']['id_user'] : null;
-        $sOnOff      = $iAutoBidOnOff === self::AUTO_BID_ON ? 'on' : 'off';
-        $sSerialized = serialize(array('id_user' => $iUserId, 'id_client' => $oClient->id_client, 'autobid_switch' => $sOnOff));
-        $oClientHistoryActions->histo(20, 'autobid_on_off', $oClient->id_client, $sSerialized);
     }
 
     /**
@@ -91,14 +68,12 @@ class AutoBidSettingsManager
     {
         /** @var \settings $oSettings */
         $oSettings = Loader::loadData('settings');
-        /** @var \client_settings $oClientSettings */
-        $oClientSettings = Loader::loadData('client_settings');
 
         $oSettings->get('Auto-bid global switch', 'type');
         $bGlobalActive = (bool)$oSettings->value;
 
         $bBetaTester = false;
-        if (false === empty($oClient->id_client) && true === (bool)$oClientSettings->getSetting($oClient->id_client, \client_setting_type::TYPE_AUTO_BID_BETA_TESTER)) {
+        if (false === empty($oClient->id_client) && true === $this->oClientSettingsManager->isBetaTester( $oClient)){
             $bBetaTester = true;
         }
 
@@ -293,4 +268,62 @@ class AutoBidSettingsManager
         }
     }
 
+
+    /**
+     * @param $iClientID
+     *
+     * @return array
+     */
+    public function getLastDateOnOff($iClientID)
+    {
+        $oClientsHistoryActions = Loader::loadData('clients_history_actions');
+        $aAutoBidHistory        = $oClientsHistoryActions->getLastAutoBidOnOffActions($iClientID);
+
+        $aDates = array();
+
+        foreach ($aAutoBidHistory as $aHistoryAction){
+            $aAction = unserialize($aHistoryAction['serialize']);
+            $aDates[$aAction['autobid_switch']] = $aHistoryAction['added'];
+        }
+        return $aDates;
+    }
+
+    /**
+     * @param \clients $oClient
+     *
+     * @return bool
+     */
+    public function isOn(\clients $oClient)
+    {
+        $oClientSettings = Loader::loadData('client_settings');
+        return (bool)$oClientSettings->getSetting($oClient->id_client, 'id_type = ' . \client_setting_type::TYPE_AUTO_BID_SWITCH . ' AND id_client');
+    }
+
+
+    /**
+     * @param $iClientId
+     * @param $sValue
+     */
+    private function saveAutoBidSwitchHistory($iClientId, $sValue)
+    {
+        /** @var \clients_history_actions $oClientHistoryActions */
+        $oClientHistoryActions = Loader::loadData('clients_history_actions');
+
+        $sOnOff      = $sValue === \client_settings::AUTO_BID_ON ? 'on' : 'off';
+        $iUserId     = isset($_SESSION['user']['id_user']) ? $_SESSION['user']['id_user'] : null;
+        $sSerialized = serialize(array('id_user' => $iUserId, 'id_client' => $iClientId, 'autobid_switch' => $sOnOff));
+        $oClientHistoryActions->histo(21, 'autobid_on_off', $iClientId, $sSerialized);
+    }
+
+    /**
+     * @param \lenders_accounts $oLenderAccount
+     * @return mixed
+     */
+
+    public function getValidationDate(\lenders_accounts $oLenderAccount){
+
+        /** @var \autobid $oAutoBid */
+        $oAutoBid = Loader::loadData('autobid');
+        return $oAutoBid->getValidationDate($oLenderAccount->id_lender_account);
+    }
 }
